Changes to hurdle calculation (#367)
* Fix hurdle percentage to use only relevant remaining votes

When the hurdle was changed to take third-party votes into account, it
would calculate the number of votes required by the trailing candidate
with an estimate of "relevant" remaining votes (i.e. excluding
third-party votes by assuming the proportion remains constant).

However, it would then divide this by the total remaining votes (i.e.
including third-party votes). This would make the trailing candidate's
hurdle percentage lower than it should be.

* Use latest data for third-party estimate when calculating hurdle

To estimate the third-party votes across the entire vote population, we
should use the largest sample size we have, which is the latest data
we've received.

This means using the latest batch to calculate hurdles for older batches
and, consequently, hurdle values for older batches might change slightly
as new data is received.

In practice, this change makes a minute difference. The largest change
for the current data was a 0.276 percentage point difference in the
proportion of third-party votes, and a ~0.1 percentage point difference
in calculated hurdle.

* Update hurdle information in ISSUE_TEMPLATE